Changed all integer references to relevant and efficient specific integer types (int to int16_t, for example). Compiles with Arduino and MBED

+++ src/arch/arduino.cpp
@@ -2,7 +2,7 @@
 #include "../common/pecan.hpp"
 #include "CAN.h"
 
-int waitPacket(CANPacket * recv_pack, int listen_id, int (*handler)(CANPacket *)){
+int16_t waitPacket(CANPacket * recv_pack, int16_t listen_id, int16_t (*handler)(CANPacket *)){
     if (recv_pack == NULL){
         CANPacket p;
         recv_pack = &p;
@@ -10,14 +10,14 @@ int waitPacket(CANPacket * recv_pack, int listen_id, int (*handler)(CANPacket *)
         // Because stack
     }
     
-    int packetsize = CAN.parsePacket();
-    int id;
+    int16_t packetsize = CAN.parsePacket();
+    int8_t id;
     if (!packetsize &&
         (id = CAN.packetId()) == listen_id
     ){
         recv_pack->id = CAN.packetId();
         recv_pack->dataSize = packetsize;
-        for (int i = 0; i < packetsize; i++){
+        for (int8_t i = 0; i < packetsize; i++){
             recv_pack->data[i] = CAN.read();
         }
         return handler(recv_pack);
@@ -26,17 +26,16 @@ int waitPacket(CANPacket * recv_pack, int listen_id, int (*handler)(CANPacket *)
 }
 
 
-int sendPacket(CANPacket * p){
+int16_t sendPacket(CANPacket * p){
     if (p->dataSize > MAX_SIZE_PACKET_DATA){
         return PACKET_TOO_BIG;
     }
     CAN.beginPacket(p->id);
-    for (int i = 0; i < p->dataSize; i++){
+    for (int8_t i = 0; i < p->dataSize; i++){
         CAN.write( p->data[i]);
     }
     if(CAN.endPacket()){
         Serial.println("Success!!!");
     }
     return SUCCESS;
-
 }
